+++ src/router.worker.ts
@@ -5,12 +5,6 @@ type FindRoute = (inputCoinAddress: string, outputCoinAddress: string, routerGra
 export let findRouteAlgos: FindRoute[];
 
 export function routerWorker(): void {
-    const addStep = (route: IRouteTvl, step: IRouteStep) => ({
-        route: [...route.route, step],
-        minTvl: Math.min(step.tvl, route.minTvl),
-        totalTvl: route.totalTvl + step.tvl,
-    });
-
     function log(fnName: string, ...args: unknown[]): void {
         if (process.env.NODE_ENV === 'development') {
             console.log(`curve-js/router-worker@${new Date().toISOString()} -> ${fnName}:`, args)
@@ -20,12 +14,12 @@ export function routerWorker(): void {
     const MAX_ROUTES_FOR_ONE_COIN = 5;
     const MAX_DEPTH = 4;
 
-    const _removeDuplications = (routes: IRouteTvl[]) =>
-        routes.filter(
-            (r, i, _routes) => _routes.map((r) => r.route.map((s) => s.poolId).toString()).indexOf(r.route.map((s) => s.poolId).toString()) === i
-        )
-    const itemAt = <T>(route: T[], index: number) => route.length > index ? route[index] : undefined;
-    const lastItem = <T>(route: T[]) => itemAt(route, route.length - 1);
+    const _removeDuplications = (routesA: IRouteTvl[], routesB: IRouteTvl[]) => {
+        const routeToStr = (r: IRouteTvl) => r.route.map((s) => s.poolId).toString();
+        const routeIdsA = new Set(routesA.map(routeToStr));
+        return routesA.concat(routesB.filter((r) => !routeIdsA.has(routeToStr(r))));
+    }
+
     const _sortByTvl = (a: IRouteTvl, b: IRouteTvl) => b.minTvl - a.minTvl || b.totalTvl - a.totalTvl || a.route.length - b.route.length;
     const _sortByLength = (a: IRouteTvl, b: IRouteTvl) => a.route.length - b.route.length || b.minTvl - a.minTvl || b.totalTvl - a.totalTvl;
 
@@ -37,14 +31,18 @@ export function routerWorker(): void {
         return swapType.toString()
     }
 
+    const _addStep = (route: IRouteTvl, step: IRouteStep) => ({
+        route: route.route.concat(step),
+        minTvl: Math.min(step.tvl, route.minTvl),
+        totalTvl: route.totalTvl + step.tvl,
+    });
+
     class SortedSizedArray<T> {
         readonly items: T[] = [];
         constructor(private readonly compareFn: (a: T, b: T) => number, private readonly maxSize: number) {}
 
         push(item: T) {
-            if (!this.fits(item)) {
-                return;
-            }
+            if (!this.fits(item)) return;
             if (this.items.length === this.maxSize) {
                 this.items.pop();
             }
@@ -55,6 +53,7 @@ export function routerWorker(): void {
                 this.items.splice(position, 0, item);
             }
         }
+
         fits(item: T): boolean {
             if (this.items.length < this.maxSize) return true;
             const last = this.items[this.items.length - 1];
@@ -67,68 +66,7 @@ export function routerWorker(): void {
 
     const _findPool = (route: IRouteTvl, poolId: string) => route.route.find((r) => r.poolId === poolId);
 
-    const _isVisitedPool = (poolId: string, route: IRouteTvl): boolean => _findPool(route, poolId) !== undefined
-
-    const _findRoutes0: FindRoute = (inputCoinAddress, outputCoinAddress, routerGraph, poolData) => {
-
-        inputCoinAddress = inputCoinAddress.toLowerCase();
-        outputCoinAddress = outputCoinAddress.toLowerCase();
-
-        const routes: IRouteTvl[] = [{ route: [], minTvl: Infinity, totalTvl: 0 }];
-        let targetRoutes: IRouteTvl[] = [];
-        let count = 0;
-        const start = Date.now();
-
-        while (routes.length > 0) {
-            count++;
-            // @ts-ignore
-            const route: IRouteTvl = routes.pop();
-            const inCoin = route.route.length > 0 ? route.route[route.route.length - 1].outputCoinAddress : inputCoinAddress;
-
-            if (inCoin === outputCoinAddress) {
-                targetRoutes.push(route);
-            } else if (route.route.length <= MAX_DEPTH) {
-                const inCoinGraph = routerGraph[inCoin];
-                for (const outCoin in inCoinGraph) {
-                    if (_isVisitedCoin(outCoin, route)) continue;
-
-                    for (const step of inCoinGraph[outCoin]) {
-                        const pool = poolData[step.poolId];
-
-                        if (!pool?.is_lending && _isVisitedPool(step.poolId, route)) continue;
-
-                        // 4 --> 6, 5 --> 7 not allowed
-                        // 4 --> 7, 5 --> 6 allowed
-                        const routePoolIdsPlusSwapType = route.route.map((s) => s.poolId + "+" + _handleSwapType(s.swapParams[2]));
-                        if (routePoolIdsPlusSwapType.includes(step.poolId + "+" + _handleSwapType(step.swapParams[2]))) continue;
-
-                        const poolCoins = pool ? pool.wrapped_coin_addresses.concat(pool.underlying_coin_addresses) : [];
-                        // Exclude such cases as:
-                        // cvxeth -> tricrypto2 -> tusd -> susd (cvxeth -> tricrypto2 -> tusd instead)
-                        if (!pool?.is_lending && poolCoins.includes(outputCoinAddress) && outCoin !== outputCoinAddress) continue;
-                        // Exclude such cases as:
-                        // aave -> aave -> 3pool (aave -> aave instead)
-                        if (pool?.is_lending && poolCoins.includes(outputCoinAddress) && outCoin !== outputCoinAddress && outCoin !== pool.token_address) continue;
-
-                        routes.push({
-                            route: [...route.route, step],
-                            minTvl: Math.min(step.tvl, route.minTvl),
-                            totalTvl: route.totalTvl + step.tvl,
-                        });
-                    }
-                }
-            }
-        }
-
-        targetRoutes = _removeDuplications([
-            ...targetRoutes.sort(_sortByTvl).slice(0, MAX_ROUTES_FOR_ONE_COIN),
-            ...targetRoutes.sort(_sortByLength).slice(0, MAX_ROUTES_FOR_ONE_COIN),
-        ]);
-        log(`[old algo] Searched ${count} routes resulting in ${targetRoutes.length} routes between ${inputCoinAddress} and ${outputCoinAddress}`, `${Date.now() - start}ms`);
-        return targetRoutes;
-    }
-
-    const _findRoutes1: FindRoute = (inputCoinAddress, outputCoinAddress, routerGraph, poolData) => {
+    const _findRoutes: FindRoute = (inputCoinAddress, outputCoinAddress, routerGraph, poolData) => {
         inputCoinAddress = inputCoinAddress.toLowerCase();
         outputCoinAddress = outputCoinAddress.toLowerCase();
 
@@ -142,7 +80,7 @@ export function routerWorker(): void {
         while (routes.length) {
             count++;
             const route = routes.pop() as IRouteTvl;
-            const inCoin = lastItem(route.route)?.outputCoinAddress ?? inputCoinAddress;
+            const inCoin = route.route.length > 0 ? route.route[route.route.length - 1].outputCoinAddress : inputCoinAddress;
             const inCoinGraph = routerGraph[inCoin];
 
             for (const outCoin in inCoinGraph) {
@@ -156,7 +94,7 @@ export function routerWorker(): void {
                         wrapped_coin_addresses = [],
                     } = poolData[step.poolId] || {};
 
-                    const currentPoolInRoute = route.route.find((r) => r.poolId === step.poolId);
+                    const currentPoolInRoute = _findPool(route, step.poolId);
                     if (currentPoolInRoute) {
                         if (!is_lending) continue;
                         // 4 --> 6, 5 --> 7 not allowed
@@ -167,7 +105,7 @@ export function routerWorker(): void {
                     }
 
                     if (step.outputCoinAddress === outputCoinAddress) {
-                        const newRoute = addStep(route, step);
+                        const newRoute = _addStep(route, step);
                         targetRoutesByTvl.push(newRoute);
                         targetRoutesByLength.push(newRoute);
                         continue;
@@ -180,7 +118,7 @@ export function routerWorker(): void {
                         if (outCoin !== token_address) continue;
                     }
                     if (route.route.length < MAX_DEPTH) {
-                        const newRoute = addStep(route, step);
+                        const newRoute = _addStep(route, step);
                         if (targetRoutesByTvl.fits(newRoute) || targetRoutesByLength.fits(newRoute)) {
                             routes.push(newRoute); // try another step
                         }
@@ -188,83 +126,16 @@ export function routerWorker(): void {
                 }
             }
         }
-        log(`[new algo, sorted array] Searched ${count} routes resulting in ${targetRoutesByTvl.items.length + targetRoutesByLength.items.length} routes between ${inputCoinAddress} and ${outputCoinAddress}`, `${Date.now() - start}ms`);
-
-        return _removeDuplications([...targetRoutesByTvl.items, ...targetRoutesByLength.items]);
-    }
-
-    const _findRoutes2: FindRoute = (inputCoinAddress, outputCoinAddress, routerGraph, poolData) => {
-        inputCoinAddress = inputCoinAddress.toLowerCase();
-        outputCoinAddress = outputCoinAddress.toLowerCase();
-
-        const routes: IRouteTvl[] = [{route: [], minTvl: Infinity, totalTvl: 0}];
-        const targetRoutes: IRouteTvl[] = [];
-
-        let count = 0;
-        const start = Date.now();
-
-        while (routes.length) {
-            count++;
-            const route = routes.pop() as IRouteTvl;
-            const inCoin = lastItem(route.route)?.outputCoinAddress ?? inputCoinAddress;
-            const inCoinGraph = routerGraph[inCoin];
-
-            for (const outCoin in inCoinGraph) {
-                if (_isVisitedCoin(outCoin, route)) continue;
-
-                for (const step of inCoinGraph[outCoin]) {
-                    const {
-                        is_lending,
-                        token_address,
-                        underlying_coin_addresses = [],
-                        wrapped_coin_addresses = [],
-                    } = poolData[step.poolId] || {};
-
-                    const currentPoolInRoute = route.route.find((r) => r.poolId === step.poolId);
-                    if (currentPoolInRoute) {
-                        if (!is_lending) continue;
-                        // 4 --> 6, 5 --> 7 not allowed
-                        // 4 --> 7, 5 --> 6 allowed
-                        if (_handleSwapType(step.swapParams[2]) === _handleSwapType(currentPoolInRoute.swapParams[2])) {
-                            continue;
-                        }
-                    }
-
-                    if (step.outputCoinAddress === outputCoinAddress) {
-                        const updatedRoute = addStep(route, step);
-                        targetRoutes.push(updatedRoute);
-                        continue;
-                    }
-
-                    if (wrapped_coin_addresses.includes(outputCoinAddress) || underlying_coin_addresses.includes(outputCoinAddress)) {
-                        // Exclude such cases as: cvxeth -> tricrypto2 -> tusd -> susd (cvxeth -> tricrypto2 -> tusd instead)
-                        if (!is_lending) continue;
-                        // Exclude such cases as: aave -> aave -> 3pool (aave -> aave instead)
-                        if (outCoin !== token_address) continue;
-                    }
-                    if (route.route.length < MAX_DEPTH) {
-                        routes.push(addStep(route, step)); // try another step
-                    }
-                }
-            }
-        }
-        log(`[new algo, normal array] Searched ${count} routes resulting in ${targetRoutes.length} routes between ${inputCoinAddress} and ${outputCoinAddress}`, `${Date.now() - start}ms`);
-
-        return _removeDuplications([
-            ...targetRoutes.sort(_sortByTvl).slice(0, MAX_ROUTES_FOR_ONE_COIN),
-            ...targetRoutes.sort(_sortByLength).slice(0, MAX_ROUTES_FOR_ONE_COIN),
-        ]);
+        return _removeDuplications(targetRoutesByTvl.items, targetRoutesByLength.items);
     }
 
     addEventListener('message', (e) => {
         const {type, routerGraph, outputCoinAddress, inputCoinAddress, poolData} = e.data;
         if (type === 'findRoutes') {
-            const routes = _findRoutes2(inputCoinAddress, outputCoinAddress, routerGraph, poolData);
+            const routes = _findRoutes(inputCoinAddress, outputCoinAddress, routerGraph, poolData);
             postMessage({type, routes});
         }
     });
-
-    findRouteAlgos = [_findRoutes0, _findRoutes1, _findRoutes2];
 }
 
 // this is a workaround to avoid importing web-worker in the main bundle (nextjs will try to inject invalid hot-reloading code)
